requirements.txt
fastapi
uvicorn
azure-identity
azure-cosmos
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv 